Fixed Amazon album fetching. The issue was how we weren't sorting our request pairs before signing, etc.
Blame neither me nor Connor nor Seth for this. This is entirely
Amazon's fault for leaving the feature poorly documented.

I forget the original functionality, but when an album is added the
add album box does not disappear, and does not give any notice that
the album has been added.